training_generic: print temperature sensor readings
Get a device structure for the SPI sensor we just added to our
devicetree. Use the sensor API to print readings.

The application will now only compile if the board's devicetree has a
"my-sensor" alias. That's not guaranteed, but nrf9160-DK has one from
the previous commit, so it still supports this app.

Concepts:

- Getting a "struct device" for a devicetree node

- Devices are generic implementations of higher level APIs,
  like "sensor"

- Drivers implement the device APIs and hide the hardware details,
  like "bme280 on SPI"

- A SPI controller is a device too. It implements the <drivers/spi.h>
  API.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>