scripts: west: improve ncs-compare heuristic
The 'west ncs-compare' extension command, among other things, will
print a list of patches that have been merged upstream which look like
they may have been taken as out of tree (OOT) patches downstream.

It currently does this by comparing shortlogs between unmerged
upstream commits and OOT downstream commits. If the edit distance
between a pair of these commits is below a configurable threshold, the
heuristic considers them potentially the same commit.

This doesn't work with restrictions that were introduced in NCS since
the tool was originally merged. Specifically, we now have sdk-zephyr
linting that limits the commit shortlog length. That has led to people
truncating upstream shortlogs that are too long to add the [nrf xyz]
sauce tag prefixes to.

Handle this case better by also checking for downstream commits that
have shortlogs which are prefixes of an upstream shortlog.

This will probably decrease the SNR of the command output for
shortlogs which are small, but that seems worth it to avoid missing
some commits.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>