samples: wifi: shell: add missing zperf option
It's not clear to me how this arose, but this sample configuration
ends up building zephyr/subsys/net/lib/zperf/zperf_tcp_receiver.c,
which in turn uses MY_IP4ADDR from zperf_internal.h as a string, which
in turn requires CONFIG_NET_CONFIG_SETTINGS=y for a non-null value
to be the resulting expansion of that string.

Turn the option on. We still have build errors in this sample due to
other renaming that happened in upstream zephyr, but it's progress:

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>